+++ sea-schema-sync/src/postgres/parser/table_constraints.rs
@@ -1,8 +1,6 @@
-use crate::Name;
 use crate::postgres::{def::*, query::TableConstraintsQueryResult};
 
-/// Assumed to be ordered by table name, then constraint name, then ordinal position, then the
-/// constraint name of the foreign key, then the ordinal position of the foreign key
+/// Assumed to be ordered by constraint name, then ordinal position.
 pub fn parse_table_constraint_query_results(
     results: Vec<TableConstraintsQueryResult>,
 ) -> Vec<Constraint> {
@@ -30,47 +28,6 @@ pub fn parse_table_constraint_query_results(
                         no_inherit: false,
                     }));
                 }
-            }
-
-            "FOREIGN KEY" => {
-                let mut columns = Vec::new();
-                let mut foreign_columns = Vec::new();
-
-                columns.push(result.column_name.unwrap());
-                let Some(table) = result.referential_key_table_name else {
-                    continue;
-                };
-                foreign_columns.push(result.referential_key_column_name.unwrap());
-                let on_update =
-                    ForeignKeyAction::from_str(&result.update_rule.clone().unwrap_or_default());
-                let on_delete =
-                    ForeignKeyAction::from_str(&result.delete_rule.clone().unwrap_or_default());
-
-                for result in results.by_ref() {
-                    if result.constraint_name != constraint_name {
-                        curr = Some(result);
-                        break;
-                    }
-
-                    if let (Some(col), Some(ref_col)) =
-                        (result.column_name, result.referential_key_column_name)
-                    {
-                        columns.push(col);
-                        foreign_columns.push(ref_col);
-                    }
-                }
-
-                columns.dedup();
-                foreign_columns.dedup();
-
-                output.push(Constraint::References(References {
-                    name: constraint_name,
-                    columns,
-                    table,
-                    foreign_columns,
-                    on_update,
-                    on_delete,
-                }));
             }
 
             "PRIMARY KEY" => {